+++ message_device_info.go
@@ -5,13 +5,13 @@ import "encoding/xml"
 type DeviceInfoQuery struct {
 	XMLName  xml.Name `xml:"Query"`
 	CmdType  string   `xml:"CmdType"`  // M, 命令类型
-	Sn       int      `xml:"SN"`       // M, sn
+	Sn       int64    `xml:"SN"`       // M, sn
 	DeviceId string   `xml:"DeviceID"` // M, 目标设备的编码
 }
 type DeviceInfoResponse struct {
 	XMLName      xml.Name `xml:"Response"`
 	CmdType      string   `xml:"CmdType"`      // M, 命令类型
-	Sn           int      `xml:"SN"`           // M, sn
+	Sn           int64    `xml:"SN"`           // M, sn
 	DeviceId     string   `xml:"DeviceID"`     // M, 目标设备的编码
 	DeviceType   string   `xml:"DeviceType"`   // O, 目标设备的名称
 	Manufacturer string   `xml:"Manufacturer"` // O, 设备生产商
