+++ message_keepalive.go
@@ -5,7 +5,7 @@ import "encoding/xml"
 type Keepalive struct {
 	XMLName  xml.Name `xml:"Notify"`
 	CmdType  string   `xml:"CmdType"`
-	Sn       int      `xml:"SN"`
+	Sn       int64    `xml:"SN"`
 	DeviceId string   `xml:"DeviceID"`
 	Status   string   `xml:"Status"`
 }
