+++ message_ptz.go
@@ -5,7 +5,7 @@ import "encoding/xml"
 type DevicePtzControl struct {
 	XMLName  xml.Name             `xml:"Control"`
 	CmdType  string               `xml:"CmdType"`  // M, 命令类型
-	Sn       int                  `xml:"SN"`       // M, sn
+	Sn       int64                  `xml:"SN"`       // M, sn
 	DeviceId string               `xml:"DeviceID"` // M, 目标设备的编码
 	PtzCmd   string               `xml:"PTZCmd"`   // M, PTZ控制命令
 	Info     DevicePtzControlInfo `xml:"Info"`     // M, 控制信息
