+++ message_record_info.go
@@ -8,7 +8,7 @@ import (
 type RecordInfoQuery struct {
 	XMLName   xml.Name  `xml:"Query"`
 	CmdType   string    `xml:"CmdType"`  // M, 命令类型
-	Sn        int       `xml:"SN"`       // M, sn
+	Sn        int64     `xml:"SN"`       // M, sn
 	DeviceId  string    `xml:"DeviceID"` // M, 目标设备的编码
 	StartTime time.Time `xml:"StartTime"`
 	EndTime   time.Time `xml:"EndTime"`
@@ -18,7 +18,7 @@ type RecordInfoQuery struct {
 type RecordInfoResponse struct {
 	XMLName    xml.Name          `xml:"Response"`
 	CmdType    string            `xml:"CmdType"`    // M, 命令类型
-	Sn         int               `xml:"SN"`         // M, sn
+	Sn         int64             `xml:"SN"`         // M, sn
 	DeviceId   string            `xml:"DeviceID"`   // M, 目标设备的编码
 	SumNum     int               `xml:"SumNum"`     // M, 总数
 	RecordList []RecordInfoEntry `xml:"RecordList"` // O, 设备列表
